05_integration/data_collection_system/requirements.txt
paho-mqtt==1.6.1
pymongo==4.6.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
